Create Hello World
This is the typical starting place for learning any programming language. You start with what people call a "Hello World" program. This program teaches you how to print text known as a "string" to the output device which is typically a terminal.

To break it down:

you have the "print()" statement
           -This prints the data between the parenthesis to the output

then we have "exit(0)"
           -This basically exits the program with a status of 0 which means that it exited successfully.
           -If we had the exit code of 1, that would mean that the program was unsuccessfully exited.