Create PasswordMaker.py
This is a program that will take a phrase or word that you give it and if long enough, make a decent password out of it. The point of this program is to give people a way of seeing that good passwords can be made from things that you can remember.