Create Tab Opener.py
This is my basic dashboard CLI used to open tabs that I want it to for automated work environment setup for class. I have another version that logs in and navigates to the correct pages as well, but I did not want to post it with the passwords. Thinking about posting a cleaned version of it.