Create textRW.py
This is a program that takes user input and saves it to a text file in the same directory called "textFile.txt". The program then reads the file and prints the contents to the console. Comments in the program can alert you to optional modifications to make it overwrite the contents of the file every time or add your input to the end of the file.