Create myClip.py
This is a program that takes allows you to easily store multiple items and copy them to your clipboard with the ease of typing a few words. All you do is type "python myClip.py <option>" and then whatever the text is that pertains to your option will be copied to your clipboard. And as an added bonus, if you want to be fast, you can follow the instructions in the comments at the end of the program to make a Win Run command for Windows for your program.