Create rate-limit-brute-force.py
This was a small program that I developed to aid me in brute forcing the login for a specific web application that was being rate limited by the IP address in the "Origin: x.x.x.x" HTTP Header. This script will basically attempt to brute force the login and then rotate the IP address in the "Origin" Header every time it encounters the rate-limiting response so that the web application thinks it is talking to a new host and the rate-limiting counter starts over.